% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{barplot.naBlocks}
\alias{barplot.naBlocks}
\title{Barplot showcasing missingness proportion of the original data}
\usage{
\method{barplot}{naBlocks}(obj, sortNA = FALSE)
}
\description{
Barplot showcasing missingness proportion of the original data
}
