% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{barplotMissProp}
\alias{barplotMissProp}
\title{Barplot Showcasing Missingness Proportion of the Original Data}
\usage{
barplotMissProp(obj, sortNA = FALSE, ...)
}
\arguments{
\item{obj}{A \code{naBlocks} object}

\item{sortNA}{A logical. If \code{FALSE} the original variable order is kept,
else the variables are ordered from least to most missingness}

\item{...}{Additional graphical arguments for \code{\link[graphics]{barplot}}}
}
\description{
Visualize the proportion of missingness per variable in a barplot.
}
\examples{
    S <- toeplitz(seq(1, 0.1, length.out = 3))
    x <- mvtnorm::rmvt(100, S, df = 5)
    y <- mice::ampute(x, mech='MCAR')$amp
    res <- classicShapeNA(y)
    barplotMissProp(res$naBlocks)
}
\seealso{
\code{\link[graphics]{barplot}}
}
