% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ellipse}
\alias{ellipse}
\title{Ellipse For Covariance Matrix}
\usage{
ellipse(mu, S, idx, n = 250)
}
\arguments{
\item{mu}{A  vector for the ellipse's center}

\item{S}{A covariance matrix}

\item{idx}{A vector of two distinct indices which specify the variables for
whose subspace a circle with mahalanobis distance 1 gets returned}

\item{n}{number of points on the ellipse}
}
\description{
Given a center and covariance, get (x,y)-coordinates of points, which have
mahalanobis distance 1.
}
\examples{
    S <- toeplitz(c(1, 0.3, 0.7))
    x <- mvtnorm::rmvt(100, S, df = 3)

    ## Plot variables 1 and 3
    idx <- c(1,3)
    plot(x[, idx])
    ## Plot projection of true covariance matrix
    lines(shapeNA:::ellipse(c(0, 0, 0), S, idx), col = 2)
}
\keyword{internal}
