% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ellipseShape}
\alias{ellipseShape}
\title{Ellipse For Covariance Matrix From M-Estimator}
\usage{
ellipseShape(obj, idx = 1:2, n = 250)
}
\arguments{
\item{obj}{A \code{shapeNA} object}

\item{idx}{A vector of two distinct indices which specify the submatrix}

\item{n}{number of points on the ellipse}
}
\description{
Given a \code{shapeNA} object, get (x,y)-coordinates of points, which for the
estimated scatter (shape, if \code{alpha} == 1) and location have mahalanobis
distance 1.
}
\examples{
    S <- toeplitz(c(1, 0.3, 0.7))
    set.seed(123)
    x <- mvtnorm::rmvt(100, S, df = 3)
    obj <- powerShape(x, alpha = 0.85)
    ## Plot variables 1 and 3
    idx <- c(1,3)
    plot(x[, idx])

    ## Plot projection of true covariance matrix
    lines(shapeNA:::ellipse(S[idx, idx], c(0, 0)), col = 2)
    ## Plot base-R estimate
    lines(shapeNA:::ellipse(cov(x[, idx]), c(0, 0)), col = 3, lty = 2)
    ## Plot M-estimate
    lines(shapeNA:::ellipseShape(obj, idx=idx), col = 4, lty = 2)
}
