% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{naBlocks}
\alias{naBlocks}
\title{Reorder Data With Missing Values}
\usage{
naBlocks(data, cleanup = TRUE, plot = FALSE)
}
\arguments{
\item{data}{A matrix with \code{NA} values}

\item{cleanup}{A logical flag. If \code{TRUE}, observations with less than 2 responses
are discarded}

\item{plot}{A logical flag. If true, a plot of the missingness pattern is produced.}
}
\value{
a \code{naBlocks} object, which is a list with \itemize{
\item \code{data} the reordered data matrix
\item \code{permutation} the permutation of the columns, which was applied to reorder
the columns according to the number of \code{NA}s
\item \code{rowPermutation} the permutation of the rows, which generates the blocks
\item \code{N} a vector of all row indices. Each row number points to the beginning
of a new missingness pattern
\item \code{D} a vector specifying the missingness pattern for each block.
\item \code{P} a vector specifying the number of observed variables per block.
\item \code{kn} a vector specifying the percentage of observed responses per variable.
}
}
\description{
Reorder a data set with \code{NA} entries to form blocks of missing values. The resulting
data will have increasing missingness along the rows and along the columns. The rows
are ordered s.t. the first block consists of complete observations, and
the following blocks are ordered from most frequent missingness pattern to least
frequent missingness pattern.
This method may fail, as it has been designed as a preprocessing step for
shape estimations.
}
