% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.naBlocks}
\alias{plot.naBlocks}
\title{plot missingness pattern of data}
\usage{
\method{plot}{naBlocks}(x, orderProp = TRUE, ...)
}
\arguments{
\item{x}{A \code{naBlocks} object}

\item{...}{additional parameters}
}
\description{
plot missingness pattern of data
}
\examples{
    x <- mvtnorm::rmvt(100, toeplitz(seq(1, 0.1, length.out = 3)), df = 5)
    y <- mice::ampute(x, mech='MCAR')$amp
    res <- classicShapeNA(y)
    plot(res$naBlocks)
}
