% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerShapeNA.R
\name{powerShapeNA}
\alias{powerShapeNA}
\alias{classicShapeNA}
\alias{tylerShapeNA}
\title{Computing M-Estimators of Shape for Data With Missing Values}
\usage{
powerShapeNA(x, alpha, center = NULL, normalization = c("det", "trace", "one"), 
         maxiter = 1e4, tol = 1e-6)

tylerShapeNA(x, center = NULL, normalization = c("det", "trace", "one"), 
          maxiter = 1e4, tol = 1e-6)

classicShapeNA(x, center = NULL, normalization = c("det", "trace", "one"),
         maxiter = 1e4, tol = 1e-6)
}
\arguments{
\item{x}{data matrix or data.frame with missing data and more than 2 columns. Representing sample from continuous distribution and MCAR missingness}

\item{alpha}{numeric, determines power function}

\item{center}{optional vector of center, if NULL the center will be estimated simultaneously to the shape estimate}

\item{normalization}{string, determines scale of returned shape estimate}

\item{maxiter}{integer, maximum number of iterations}

\item{tol}{numeric, tolerance level}
}
\value{
shape and center estimate
}
\description{
todo: add description here
}
\examples{
    ## generate data set with missing values
    x <- mvtnorm::rmvt(100, toeplitz(seq(1, 0.1, length.out = 3)), df = 5)
    y <- mice::ampute(x, mech='MCAR')$amp
    ## compute M-estimate
    res <- powerShapeNA(y, alpha = 0.5)
    summary(res)
}
\references{
Frahm, G., & Jaekel, U. (2010). A generalization of Tyler’s M-estimators to the case of incomplete data. Computational Statistics & Data Analysis, 54, 374-393. \url{doi:10.1016/j.csda.2009.08.019}.

Frahm, G., Nordhausen, K., & Oja, H. (2020). M-estimation with incomplete and dependent multivariate data. Journal of Multivariate Analysis, 176, 104569. \url{doi:10.1016/j.jmva.2019.104569}.
}
