% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.shapeNA}
\alias{print.shapeNA}
\title{print method for elements of class shapeNA}
\usage{
\method{print}{shapeNA}(obj)
}
\description{
Only print M-estimates and alpha level
}
\examples{
    x <- mvtnorm::rmvt(100, toeplitz(seq(1, 0.1, length.out = 3)), df = 5)
    res <- tylerShape(x)
    res ## equivalent to call print(res)
}
