% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.summary.shapeNA}
\alias{print.summary.shapeNA}
\title{Print Method For Class \code{summary.shapeNA}}
\usage{
\method{print}{summary.shapeNA}(x, ...)
}
\arguments{
\item{x}{object returned from summary.shapeNA}

\item{...}{further arguments}
}
\value{
invisibly return NULL
}
\description{
Print Method For Class \code{summary.shapeNA}
}
\examples{
    obj <- tylerShape(mvtnorm::rmvt(100, diag(3)))
    print(summary(obj))
}
