% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ellipse}
\alias{ellipse}
\title{Ellipse For Covariance Matrix}
\usage{
ellipse(mu, S, n = 250)
}
\arguments{
\item{mu}{A 2-dimensional vector for the ellipse's center}

\item{S}{A 2x2 covariance matrix}

\item{n}{number of points on the ellipse}
}
\description{
Given a center and covariance, get (x,y)-coordinates of points, which have
mahalanobis distance 1.
}
\examples{
    S <- toeplitz(c(1, 0.3, 0.7))
    x <- mvtnorm::rmvt(100, S, df = 3)

    ## Plot variables 1 and 3
    idx <- c(1,3)
    plot(x[, idx])
    ## Plot projection of true covariance matrix
    lines(shapeNA:::ellipse(c(0, 0), S[idx, idx]), col = 2)
}
