% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{naBlocks}
\alias{naBlocks}
\title{Reorder Data With Missing Values}
\usage{
naBlocks(x, cleanup = TRUE, plot = FALSE)
}
\arguments{
\item{cleanup}{A logical flag. If \code{TRUE}, observations with less than 2 responses
are discarded.}

\item{plot}{A logical flag. If true, a plot of the missingness pattern is produced.}

\item{data}{A matrix with \code{NA} values.}
}
\description{
Reorder a data set with \code{NA} entries to form blocks of missing values. The resulting
data will have increasing missingness along the rows and along the columns. The rows
are ordered such that the first block consists of complete observations, and
the following blocks are ordered from most frequent missingness pattern to least
frequent missingness pattern.
This method may fail if the missingness is too strong or if the number of
observations is too low (the number of observations has to exceed the number
of variables), as it has been designed as a preprocessing step for shape estimations.
}
