% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.shapeNA}
\alias{plot.shapeNA}
\title{Crude visualization of shape estimate}
\usage{
\method{plot}{shapeNA}(obj, legend = TRUE, message = TRUE)
}
\description{
If estimate comes from missing data, additionally the columns are
marked with a colored bar, indicating their missingness proportion
}
