% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.shapeNA}
\alias{plot.shapeNA}
\title{Visualization of Shape Estimate}
\usage{
\method{plot}{shapeNA}(x, message = TRUE, ...)
}
\arguments{
\item{x}{A \code{shapeNA} oopbject}

\item{message}{A logical, If \code{TRUE}, the percentage of observed values per
variable is printed in the console.}

\item{...}{Additional parameters.}
}
\description{
Plot each matrix entry as a cell, with dark cells indicating high values and
light values indicate small values.
}
\examples{
    x <- mvtnorm::rmvt(100, toeplitz(seq(1, 0.1, length.out = 3)), df = 5)
    y <- mice::ampute(x, mech='MCAR')$amp
    res <- tylerShapeNA(y)
    plot(res)
}
