% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerShape.R
\name{powerShape}
\alias{powerShape}
\alias{tylerShape}
\alias{classicShape}
\title{Compute shape estimate for full data}
\usage{
powerShape(x, alpha, center = NULL, normalization = c("det", "trace", "one"), maxiter = 1e4, tol = 1e-6)

tylerShape(x, center = NULL, normalization = c("det", "trace", "one"), maxiter = 1e4, tol = 1e-6)

classicShape(x, center = NULL, normalization = c("det", "trace", "one"), maxiter = 1e4, tol = 1e-6)
}
\arguments{
\item{x}{numeric data matrix or data.frame without missing data. Representing sample from continuous distribution}

\item{alpha}{numeric, determines power function}

\item{center}{optional vector of center, if NULL the center will be estimated simultaneously to the shape estimate}

\item{normalization}{string, determines scale of returned shape estimate}

\item{maxiter}{integer, maximum number of itreations}

\item{tol}{numeric, tolerance level}
}
\value{
a shapeNA object, which contains a shape and center estimate
}
\description{
Given a data matrix \code{x} from a continuous distribution, return a shape estimate and, if not supplied as \code{center}, a location estimate.
}
\examples{
    x <- mvtnorm::rmvt(100, toeplitz(seq(1, 0.1, length.out=5)))
    res <- powerShape(x, alpha=0.67, normalization='one')

}
\references{
Frahm, G., & Jaekel, U. (2010). A generalization of Tyler’s M-estimators to the case of incomplete data. Computational Statistics & Data Analysis, 54(2), 374-393.

Frahm, G., Nordhausen, K., & Oja, H. (2020). M-estimation with incomplete and dependent multivariate data. Journal of Multivariate Analysis, 176, 104569.
}
\seealso{
powerShapeNA

tylerShapeNA

classicShapeNA
}
