% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.shapeNA}
\alias{print.shapeNA}
\title{Print Method for Elements of Class \code{shapeNA}}
\usage{
\method{print}{shapeNA}(x, ...)
}
\arguments{
\item{x}{A \code{shapeNA} object}

\item{...}{Additional parameters.}
}
\description{
Print alpha level, shape estimate and center.
}
\examples{
    x <- mvtnorm::rmvt(100, toeplitz(seq(1, 0.1, length.out = 3)), df = 5)
    res <- tylerShape(x)
    res ## equivalent to call print(res)
}
