% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.shapeNA}
\alias{print.shapeNA}
\title{Print Method for Objects of Class \code{shapeNA}}
\usage{
\method{print}{shapeNA}(x, ...)
}
\arguments{
\item{x}{A \code{shapeNA} object}

\item{...}{Additional parameters passed to lower level \code{\link[base]{print}}.}
}
\value{
No return value.
}
\description{
Prints the chosen value of \code{alpha} as well as the estimated shape and
location for objects of class \code{shapeNA}.
}
\examples{
    x <- mvtnorm::rmvt(100, toeplitz(seq(1, 0.1, length.out = 3)), df = 5)
    res <- tylerShape(x)
    res ## equivalent to call print(res)
}
