% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.summary.shapeNA}
\alias{print.summary.shapeNA}
\title{print method for class summary.shapeNA}
\usage{
\method{print}{summary.shapeNA}(obj, ...)
}
\arguments{
\item{obj}{object returned from summary.shapeNA}

\item{...}{further arguments}
}
\value{
invisibly return NULL
}
\description{
print method for class summary.shapeNA
}
\examples{
    obj <- tylerShape(mvtnorm::rmvt(100, diag(3)))
    print(summary(obj))
}
