% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.summary.shapeNA}
\alias{print.summary.shapeNA}
\title{Print Method for Class \code{summary.shapeNA}}
\usage{
\method{print}{summary.shapeNA}(x, ...)
}
\arguments{
\item{x}{Object returned from \code{\link{summary.shapeNA}}.}

\item{...}{Further arguments to be passed to or from methods.}
}
\value{
No return value.
}
\description{
Print Method for Class \code{summary.shapeNA}
}
\examples{
    obj <- tylerShape(mvtnorm::rmvt(100, diag(3)))
    print(summary(obj))
}
