% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.shapeNA}
\alias{summary.shapeNA}
\title{Summary Method for Class \code{shapeNA}}
\usage{
\method{summary}{shapeNA}(object, ...)
}
\arguments{
\item{object}{an object of class shapeNA, usually from a call to powerShape or similar functions}

\item{...}{further arguments}
}
\description{
Summary Method for Class \code{shapeNA}
}
\examples{
    obj <- tylerShape(mvtnorm::rmvt(100, diag(3)))
    summary(obj)
}
