% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.shapeNA}
\alias{summary.shapeNA}
\title{Summary Method for Class \code{shapeNA}}
\usage{
\method{summary}{shapeNA}(object, ...)
}
\arguments{
\item{object}{An object of class \code{shapeNA}, usually from a call to
\code{\link{powerShape}} or similar functions.}

\item{...}{Further arguments to be passed to or from methods.}
}
\value{
A \code{summary.shapeNA} object. For objects of this class, the \code{print}
method tries to format the location and shape estimate in a readable format
and also shows the number of iterations, before the algorithm converged.
}
\description{
Summary methods for objects from class \code{shapeNA}.
}
\examples{
    obj <- tylerShape(mvtnorm::rmvt(100, diag(3)))
    summary(obj)
}
